# Conflicts:
#	backend/build.gradle
#	backend/src/main/resources/application.yaml

+++ backend/src/main/java/com/paradise/event_ticket_system/notification/confirmation/api/PurchaseConfirmationResponse.java
@@ -2,7 +2,7 @@ package com.paradise.event_ticket_system.notification.confirmation.api;
 
 public record PurchaseConfirmationResponse(
 	Long deliveryId,
-	Integer orderId,
+	Long orderId,
 	String orderReference,
 	PurchaseConfirmationOutcome status,
 	boolean alreadyProcessed
