+++ backend/src/main/java/com/paradise/event_ticket_system/notification/confirmation/domain/PurchaseConfirmationDeliveryRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface PurchaseConfirmationDeliveryRepository
 	extends JpaRepository<PurchaseConfirmationDelivery, Long> {
 
-	Optional<PurchaseConfirmationDelivery> findByOrderId(Integer orderId);
+	Optional<PurchaseConfirmationDelivery> findByOrderId(Long orderId);
 }
